+++ PostHog/Utils/AssociatedKeys.swift
@@ -1,0 +1,14 @@
+//
+//  AssociatedKeys.swift
+//  PostHog
+//
+//  Created by Yiannis Josephides on 04/12/2024.
+//
+
+import Foundation
+
+enum AssociatedKeys {
+    static var phForwardingDelegate: UInt8 = 0
+    static var phNoCapture: UInt8 = 0
+    static var phLabel: UInt8 = 0
+}