+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: ability to add a custom label to autocapture elements ([#271](https://github.com/PostHog/posthog-ios/pull/271))
+
 ## 3.16.2 - 2024-12-05
 
 - fix: ignore autocapture events from keyboard window ([#269](https://github.com/PostHog/posthog-ios/pull/269))